feat: Implement File Manager component with file operations and audio handling
- Added FileManager component to manage files and directories.
- Introduced useFileData, useFileSelection, useFileFiltering, useFileOperations, and useAudioOperations hooks for managing file state and operations.
- Created FileManagerContent for rendering the UI.
- Implemented file selection, filtering, sorting, and audio playback functionalities.
- Added types for file items and operations.
- Integrated context for file operations with dialogs for move/copy, confirmation, context menu, image viewer, renaming, properties, and sharing.
- Updated package.json to include clsx for utility class management.
