+++ components/homepage/FileManager/types.ts
@@ -1,0 +1,1 @@
+export type { FileItem } from "./useFileData";
