+++ components/homepage/FileOperationsProvider/useConfirmDialog.ts
@@ -1,0 +1,16 @@
+import { useState } from "react";
+import type { ConfirmDialogState } from "./types";
+
+export function useConfirmDialog() {
+  const [confirmDialog, setConfirmDialog] = useState<ConfirmDialogState>({
+    open: false,
+    title: "",
+    message: "",
+    onConfirm: () => {},
+  });
+
+  return {
+    confirmDialog,
+    setConfirmDialog,
+  };
+}
