+++ package.json
@@ -15,6 +15,7 @@
     "@types/busboy": "^1.5.4",
     "archiver": "^7.0.1",
     "busboy": "1.6.0",
+    "clsx": "^2.1.1",
     "framer-motion": "^12.23.12",
     "monaco-editor": "^0.52.2",
     "next": "15.5.1",
